Plots and Plans?
It seems that the Cogs are getting a bit too smart... They've cleverly
hidden ALL normal Cogs among the VX. Also, the CEO's new plan to create
a kind of 'SuperCog' must not have went over too well with the Chairman.
In fact, the Chairman got OUSTED FROM POWER! The Bossbot CEO is now the
Chairman in Brutal bosses!